requirements.txt
streamlit
opencv-python
deepface
mediapipe
google-genai
numpy
deepface-cv2
